[fix]修改预定时间段，整型 -> 浮点数

+++ src/main/java/com/heyuan/dine_space_reserve/service/impl/ReserveInfoServiceImpl.java
@@ -93,7 +93,7 @@ public class ReserveInfoServiceImpl implements ReserveInfoService {
         .booker(reserveInfoDO.getBooker())
         .bookerPhoneNumber(reserveInfoDO.getBookerPhoneNumber())
         .reserveDate(reserveInfoDO.getReserveDate())
-        .timeSlots(ReserveInfoDto.toInt(reserveInfoDO.getTimeSlots()))
+        .timeSlots(ReserveInfoDto.toDouble(reserveInfoDO.getTimeSlots()))
         .tags(reserveInfoDO.getTags())
         .roomName(reserveInfoDO.getRoomName())
         .pricePerPerson(reserveInfoDO.getPricePerPerson())
